pyproject.toml
[project]
name = "cogcat"
version = "0.1.0"
description = "Terminal GeoTIFF previewer using half-block ANSI characters"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9",
    "rich>=13",
    "rich-pixels>=3",
    "rasterio>=1.3",
    "Pillow>=10",
    "numpy>=1.24",
]

[project.optional-dependencies]
colormaps = ["matplotlib>=3.7"]

[project.scripts]
cogcat = "cogcat.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cogcat"]
